+++ Interview_Tracking_System/src/main/java/com/spring/json/ITS_TBL_Interview_Schedule.java
@@ -8,16 +8,16 @@ public class ITS_TBL_Interview_Schedule {
 	private String subject;
 	private LocalDate interviewDate;
 	private LocalTime interviewTime;
-	private Float techRating;
+	private Integer techRating;
 	private LocalDate empHRInterviewDate;
 	private LocalDate empHRInterviewTime;
-	private Float empHRRating;
+	private Integer empHRRating;
 	private String result;
-	private int shareResult;
+	private Integer shareResult;
 	
 	public ITS_TBL_Interview_Schedule(long interviewId, String subject, LocalDate interviewDate,
-		LocalTime interviewTime, Float techRating, LocalDate empHRInterviewDate, LocalDate empHRInterviewTime,
-		Float empHRRating, String result, int shareResult) {
+		LocalTime interviewTime, Integer techRating, LocalDate empHRInterviewDate, LocalDate empHRInterviewTime,
+		Integer empHRRating, String result, Integer shareResult) {
 		super();
 		this.interviewId = interviewId;
 		this.subject = subject;
@@ -31,12 +31,12 @@ public class ITS_TBL_Interview_Schedule {
 		this.shareResult = shareResult;
 	}
 	
-	public ITS_TBL_Interview_Schedule(Float empHRRating) {
+	public ITS_TBL_Interview_Schedule(Integer empHRRating) {
 		super();
 		this.empHRRating = empHRRating;
 	}
 	
-	public ITS_TBL_Interview_Schedule(String subject, Float techRating) {
+	public ITS_TBL_Interview_Schedule(String subject, Integer techRating) {
 		super();
 		this.subject = subject;
 		this.techRating = techRating;
@@ -70,10 +70,10 @@ public class ITS_TBL_Interview_Schedule {
 	public void setInterviewTime(LocalTime interviewTime) {
 		this.interviewTime = interviewTime;
 	}
-	public Float getTechRating() {
+	public Integer getTechRating() {
 		return techRating;
 	}
-	public void setTechRating(Float techRating) {
+	public void setTechRating(Integer techRating) {
 		this.techRating = techRating;
 	}
 	public LocalDate getEmpHRInterviewDate() {
@@ -88,10 +88,10 @@ public class ITS_TBL_Interview_Schedule {
 	public void setEmpHRInterviewTime(LocalDate empHRInterviewTime) {
 		this.empHRInterviewTime = empHRInterviewTime;
 	}
-	public Float getEmpHRRating() {
+	public Integer getEmpHRRating() {
 		return empHRRating;
 	}
-	public void setEmpHRRating(Float empHRRating) {
+	public void setEmpHRRating(Integer empHRRating) {
 		this.empHRRating = empHRRating;
 	}
 	public String getResult() {
@@ -100,10 +100,10 @@ public class ITS_TBL_Interview_Schedule {
 	public void setResult(String result) {
 		this.result = result;
 	}
-	public int getShareResult() {
+	public Integer getShareResult() {
 		return shareResult;
 	}
-	public void setShareResult(int shareResult) {
+	public void setShareResult(Integer shareResult) {
 		this.shareResult = shareResult;
 	}
